+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+Version 0.1.6
+=============
+
+- Implemented `upload` command
+- Added `download` command
+- Added `pbstemp` command
+
 Version 0.1.5
 =============